backend/requirements.txt
# 共享菜园云端小筑 - 后端依赖包

# Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# 数据库
sqlalchemy==2.0.25
pymysql==1.1.0
redis==5.0.1

# 认证与安全
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# 微信相关
requests==2.31.0

# 工具库
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# 日期时间
python-dateutil==2.8.2

# CORS
fastapi-cors==0.0.6

# 定时任务
apscheduler==3.10.4
